Move addresslist state to the contract from the package
