+++ contracts/andromeda_addresslist/src/lib.rs
@@ -1,3 +1,4 @@
 pub mod contract;
+mod state;
 #[cfg(test)]
 mod testing;
